pyproject.toml
[project]
name = "airflow-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "airflow_pipeline/README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyterlab>=4.5.6",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.35.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.17",
    "psycopg2-binary>=2.9.0",
]
